services/api/requirements.txt
# API service dependencies
manga-shared @ file:../../packages/shared
fastapi
uvicorn
pydantic==2.5.0
python-multipart
python-dotenv
redis[hiredis]
requests
supabase
omise
aiohttp
Pillow
httpx==0.27.2
sentry-sdk[fastapi]
openai
# Needed by manga_translator imports (config, utils, to_json)
colorama
omegaconf
opencv-python-headless
numpy==1.26.4
einops
shapely
tqdm
pyclipper
freetype-py
marshmallow
py3langid==0.2.2
